+++ ProjectFinal/src/logic/Bill.java
@@ -55,4 +55,5 @@ public class Bill {
 		return id;
 	}
 	
+	
 }
